fix: payment service & reservation service & point service 테스트
